fdc: fixed dataIndex to be uint16_t

+++ include/Drive/FDC177x.h
@@ -125,7 +125,7 @@ class FDC177x : public DriveFDCBase
 
         uint8_t sectorBuffer[MaxSectorSize]{};
         uint16_t currentSectorSize;
-        uint8_t dataIndex;
+        uint16_t dataIndex;
         bool readSectorInProgress;
         bool writeSectorInProgress;
